requirements-dev.txt
# TESTS
coverage==3.7.1
django-coverage==1.2.4
nose==1.3.0
django-nose==1.2
nose-progressive
tox==1.7.0

coveralls
python-coveralls

flake8==2.1.0

django-dynamic-fixture

# DOCS
sphinx
sphinx_rtd_theme
